password generation functions moved to separate file in default_auth plugin
